feat: editTopic layout
